Extract fragment caching methods into concern
No specs moved as this code seems to be testing only partial when used
to get attachments in the RequestController.
